refactor(patina): move reserved-name extraction helpers into submodule
