+++ crates/vize_patina/src/rules/vue/no_reserved_component_names.rs
@@ -27,19 +27,17 @@
 //! <template><div /></template>
 //! ```
 
+use self::extract::{define_options_name, find_component_options, name_string_literal};
 use crate::context::LintContext;
 use crate::diagnostic::Severity;
 use crate::ir::ByteRange;
 use crate::rule::{Rule, RuleCategory, RuleMeta};
 use crate::rules::script::script_source_type;
 use oxc_allocator::Allocator as OxcAllocator;
-use oxc_ast::ast::{
-    Argument, BindingPattern, CallExpression, ExportDefaultDeclarationKind, Expression,
-    ObjectExpression, ObjectPropertyKind, Program, PropertyKey, Statement, StringLiteral,
-};
+use oxc_ast::ast::StringLiteral;
 use oxc_parser::Parser;
+use vize_carton::String;
 use vize_carton::is_html_tag;
-use vize_carton::{FxHashMap, String};
 use vize_croquis::builtins::is_builtin_component;
 
 static META: RuleMeta = RuleMeta {
@@ -213,150 +211,7 @@ struct ComponentNameFinding {
     end: u32,
 }
 
-fn name_string_literal<'a>(options: &'a ObjectExpression<'a>) -> Option<&'a StringLiteral<'a>> {
-    for property in &options.properties {
-        let ObjectPropertyKind::ObjectProperty(property) = property else {
-            continue;
-        };
-        if property.computed {
-            continue;
-        }
-        if !matches!(property_key_name(&property.key), Some("name")) {
-            continue;
-        }
-        if let Expression::StringLiteral(literal) = &property.value {
-            return Some(literal);
-        }
-    }
-    None
-}
-
-fn define_options_name<'a>(program: &'a Program<'a>) -> Option<&'a StringLiteral<'a>> {
-    for statement in program.body.iter() {
-        let Statement::ExpressionStatement(expression) = statement else {
-            continue;
-        };
-        let Expression::CallExpression(call) = &expression.expression else {
-            continue;
-        };
-        let Expression::Identifier(callee) = &call.callee else {
-            continue;
-        };
-        if !matches!(callee.name.as_str(), "defineOptions") {
-            continue;
-        }
-        if let Some(Argument::ObjectExpression(object)) = call.arguments.first() {
-            return name_string_literal(object);
-        }
-    }
-    None
-}
-
-fn property_key_name<'a>(key: &'a PropertyKey<'a>) -> Option<&'a str> {
-    match key {
-        PropertyKey::StaticIdentifier(id) => Some(id.name.as_str()),
-        PropertyKey::StringLiteral(string) => Some(string.value.as_str()),
-        _ => None,
-    }
-}
-
-fn find_component_options<'a>(program: &'a Program<'a>) -> Option<&'a ObjectExpression<'a>> {
-    let mut bindings: FxHashMap<&'a str, &'a ObjectExpression<'a>> = FxHashMap::default();
-
-    for statement in program.body.iter() {
-        let Statement::VariableDeclaration(declaration) = statement else {
-            continue;
-        };
-        for declarator in &declaration.declarations {
-            let Some(init) = declarator.init.as_ref() else {
-                continue;
-            };
-            if let BindingPattern::BindingIdentifier(id) = &declarator.id
-                && let Some(object) = options_from_expression(init, &bindings)
-            {
-                bindings.insert(id.name.as_str(), object);
-            }
-        }
-    }
-
-    for statement in program.body.iter() {
-        let Statement::ExportDefaultDeclaration(export) = statement else {
-            continue;
-        };
-        if let Some(object) = options_from_export(&export.declaration, &bindings) {
-            return Some(object);
-        }
-    }
-
-    None
-}
-
-fn options_from_export<'a>(
-    declaration: &'a ExportDefaultDeclarationKind<'a>,
-    bindings: &FxHashMap<&'a str, &'a ObjectExpression<'a>>,
-) -> Option<&'a ObjectExpression<'a>> {
-    match declaration {
-        ExportDefaultDeclarationKind::ObjectExpression(object) => Some(object),
-        ExportDefaultDeclarationKind::CallExpression(call) => options_from_call(call, bindings),
-        ExportDefaultDeclarationKind::Identifier(identifier) => {
-            bindings.get(identifier.name.as_str()).copied()
-        }
-        ExportDefaultDeclarationKind::ParenthesizedExpression(paren) => {
-            options_from_expression(&paren.expression, bindings)
-        }
-        ExportDefaultDeclarationKind::TSAsExpression(ts_as) => {
-            options_from_expression(&ts_as.expression, bindings)
-        }
-        ExportDefaultDeclarationKind::TSSatisfiesExpression(ts_satisfies) => {
-            options_from_expression(&ts_satisfies.expression, bindings)
-        }
-        ExportDefaultDeclarationKind::TSNonNullExpression(ts_non_null) => {
-            options_from_expression(&ts_non_null.expression, bindings)
-        }
-        _ => None,
-    }
-}
-
-fn options_from_expression<'a>(
-    expression: &'a Expression<'a>,
-    bindings: &FxHashMap<&'a str, &'a ObjectExpression<'a>>,
-) -> Option<&'a ObjectExpression<'a>> {
-    match expression {
-        Expression::ObjectExpression(object) => Some(object),
-        Expression::CallExpression(call) => options_from_call(call, bindings),
-        Expression::Identifier(identifier) => bindings.get(identifier.name.as_str()).copied(),
-        Expression::ParenthesizedExpression(paren) => {
-            options_from_expression(&paren.expression, bindings)
-        }
-        Expression::TSAsExpression(ts_as) => options_from_expression(&ts_as.expression, bindings),
-        Expression::TSSatisfiesExpression(ts_satisfies) => {
-            options_from_expression(&ts_satisfies.expression, bindings)
-        }
-        Expression::TSNonNullExpression(ts_non_null) => {
-            options_from_expression(&ts_non_null.expression, bindings)
-        }
-        _ => None,
-    }
-}
-
-fn options_from_call<'a>(
-    call: &'a CallExpression<'a>,
-    bindings: &FxHashMap<&'a str, &'a ObjectExpression<'a>>,
-) -> Option<&'a ObjectExpression<'a>> {
-    let Expression::Identifier(callee) = &call.callee else {
-        return None;
-    };
-    if !matches!(callee.name.as_str(), "defineComponent" | "_defineComponent") {
-        return None;
-    }
-    match call.arguments.first()? {
-        Argument::ObjectExpression(object) => Some(object),
-        Argument::Identifier(identifier) => bindings.get(identifier.name.as_str()).copied(),
-        argument => argument
-            .as_expression()
-            .and_then(|expression| options_from_expression(expression, bindings)),
-    }
-}
+mod extract;
 
 #[cfg(test)]
 mod tests;
